+++ crates/tycho-substreams/src/lib.rs
@@ -1,4 +1,9 @@
 pub mod balances;
 pub mod contract;
 mod mock_store;
-pub mod pb;
+pub mod models;
+mod pb;
+
+pub mod prelude {
+    pub use super::models::*;
+}
